requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asttokens==3.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
dataclasses-json==0.6.7
ddgs==9.6.0
debugpy==1.8.17
decorator==5.2.1
duckduckgo-search==8.1.1
executing==2.2.1
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.7.0
ipykernel==6.30.1
ipython==9.6.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-client==8.6.3
jupyter-core==5.8.1
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.79
langchain-ollama==0.3.10
langchain-text-splitters==0.3.11
langgraph==0.6.10
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.34
lxml==6.0.2
markdown-it-py==4.0.0
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.7.0
mypy-extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.3
ollama==0.6.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.0
primp==0.15.0
prompt-toolkit==3.0.52
propcache==0.4.1
psutil==7.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.12.0
pydantic-core==2.41.1
pydantic-settings==2.11.0
pyee==11.1.1
pygments==2.19.2
pyppeteer==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyyaml==6.0.3
pyzmq==27.1.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
sqlalchemy==2.0.44
stack-data==0.6.3
tenacity==9.1.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==1.26.20
wcwidth==0.2.14
websockets==10.4
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
